+++ src/Commands/Utilities/BrandCenterUtility.cs
@@ -1,3 +1,4 @@
+using Microsoft.Office.SharePoint.Tools;
 using Microsoft.SharePoint.Client;
 using PnP.PowerShell.Commands.Base;
 using PnP.PowerShell.Commands.Model.SharePoint.BrandCenter;
@@ -23,7 +24,7 @@ namespace PnP.PowerShell.Commands.Utilities
         /// <param name="store">The store to check for the font. When NULL, it will check all stores.</param>
         /// <param name="clientContext">ClientContext to use to communicate with SharePoint Online</param>
         /// <returns>The font with the provided identity or NULL if no font found with the provided identity</returns>
-        public static Font GetFontPackageByTitle(BasePSCmdlet cmdlet, ClientContext clientContext, string title, string webUrl, Store store = Store.All)
+        public static FontPackage GetFontPackageByTitle(BasePSCmdlet cmdlet, ClientContext clientContext, string title, string webUrl, Store store = Store.All)
         {
             if (store == Store.All)
             {
@@ -36,7 +37,7 @@ namespace PnP.PowerShell.Commands.Utilities
             cmdlet?.LogDebug($"Making a GET request to {url} to retrieve {store} font with title {title}.");
             try
             {
-                var font = RestHelper.Get<Font>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext);
+                var font = RestHelper.Get<FontPackage>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext);
 
                 if (font != null)
                 {
@@ -60,7 +61,7 @@ namespace PnP.PowerShell.Commands.Utilities
         /// <param name="store">The store to check for the font. When NULL, it will check all stores.</param>
         /// <param name="clientContext">ClientContext to use to communicate with SharePoint Online</param>
         /// <returns>The font with the provided identity or NULL if no font found with the provided identity</returns>
-        public static Font GetFontPackageById(BasePSCmdlet cmdlet, ClientContext clientContext, Guid identity, string webUrl, Store store = Store.All)
+        public static FontPackage GetFontPackageById(BasePSCmdlet cmdlet, ClientContext clientContext, Guid identity, string webUrl, Store store = Store.All)
         {
             if (store == Store.All)
             {
@@ -73,7 +74,7 @@ namespace PnP.PowerShell.Commands.Utilities
             cmdlet?.LogDebug($"Making a GET request to {url} to retrieve {store} font with identity {identity}.");
             try
             {
-                var font = RestHelper.Get<Font>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext);
+                var font = RestHelper.Get<FontPackage>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext);
 
                 if (font != null)
                 {
@@ -96,11 +97,11 @@ namespace PnP.PowerShell.Commands.Utilities
         /// <param name="store">The store to check for the font. When NULL, it will check all stores.</param>
         /// <param name="clientContext">ClientContext to use to communicate with SharePoint Online</param>
         /// <returns>The available fonts</returns>
-        public static List<Font> GetFontPackages(BasePSCmdlet cmdlet, ClientContext clientContext, string webUrl, Store store = Store.All)
+        public static List<FontPackage> GetFontPackages(BasePSCmdlet cmdlet, ClientContext clientContext, string webUrl, Store store = Store.All)
         {
             if (store == Store.All)
             {
-                var allStoresFonts = new List<Font>();
+                var allStoresFonts = new List<FontPackage>();
                 allStoresFonts.AddRange(GetFontPackages(cmdlet, clientContext, webUrl, Store.Site));
                 allStoresFonts.AddRange(GetFontPackages(cmdlet, clientContext, webUrl, Store.Tenant));
                 allStoresFonts.AddRange(GetFontPackages(cmdlet, clientContext, webUrl, Store.OutOfBox));
@@ -109,7 +110,7 @@ namespace PnP.PowerShell.Commands.Utilities
 
             var url = GetStoreFontPackageUrlByStoreType(store, webUrl);
             cmdlet?.LogDebug($"Making a GET request to {url} to retrieve {store} fonts.");
-            var fonts = RestHelper.Get<RestResultCollection<Font>>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext);
+            var fonts = RestHelper.Get<RestResultCollection<FontPackage>>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext);
             return fonts.Items.ToList();
         }
 
@@ -122,7 +123,8 @@ namespace PnP.PowerShell.Commands.Utilities
         /// <param name="title">Title of the font package to add</param>
         /// <param name="visble">Indicates if the font package should be visible in the Brand Center. Defaults to true.</param>
         /// <param name="clientContext">ClientContext to use to communicate with SharePoint Online</param>
-        public static void AddFontPackage(BasePSCmdlet cmdlet, ClientContext clientContext, Store store, string webUrl, string title, bool visble = true)
+        /// <returns>The created FontPackage instance</returns>
+        public static FontPackage AddFontPackage(BasePSCmdlet cmdlet, ClientContext clientContext, Store store, string webUrl, string title, Font displayFont, Font contentFont, Font titleFont, string titleFontStyle, Font headlineFont, string headlineFontStyle, Font bodyFont, string bodyFontStyle, Font interactiveFont, string interactiveFontStyle, bool visble = true)
         {
             if (store != Store.Tenant && store != Store.Site)
             {
@@ -132,10 +134,33 @@ namespace PnP.PowerShell.Commands.Utilities
             {
                 throw new ArgumentNullException(nameof(webUrl), "Web URL cannot be null or empty.");
             }
- 
+
+            // Validate that the provided fonts are all either the displayFont or contentFont
+            if (bodyFont.Id != displayFont.Id && bodyFont.Id != contentFont.Id)
+            {
+                throw new ArgumentException($"{nameof(bodyFont)} must be either the {nameof(bodyFont)} or {nameof(contentFont)} font.");
+            }
+            if (headlineFont.Id != displayFont.Id && headlineFont.Id != contentFont.Id)
+            {
+                throw new ArgumentException($"{nameof(headlineFont)} must be either the {nameof(displayFont)} or {nameof(contentFont)} font.");
+            }
+            if (interactiveFont.Id != displayFont.Id && interactiveFont.Id != contentFont.Id)
+            {
+                throw new ArgumentException($"{nameof(interactiveFont)} must be either the {nameof(displayFont)} or {nameof(contentFont)} font.");
+            }
+            if (titleFont.Id != displayFont.Id && titleFont.Id != contentFont.Id)
+            {
+                throw new ArgumentException($"{nameof(titleFont)} must be either the {nameof(displayFont)} or {nameof(contentFont)} font.");
+            }
+
+            // Generate unique IDs for the fonts to avoid conflicts
+            var displayFontUniqueId = $"{displayFont.Name}-{new Random().NextInt64(1000000000L, 9999999999L)}";
+            var contentFontUniqueId = $"{contentFont.Name}-{new Random().NextInt64(1000000000L, 9999999999L)}";
+
+            // Create the font package
             var url = GetStoreFontPackageUrlByStoreType(store, webUrl);
             cmdlet?.LogDebug($"Making a POST request to {url} to create a {store} Brand Center font package.");
-            var response = RestHelper.Post<Font>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext, new Font
+            var fontpackage = RestHelper.Post<FontPackage>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext, new FontPackage
             {
                 Title = title,
                 Store = store,
@@ -143,20 +168,22 @@ namespace PnP.PowerShell.Commands.Utilities
                 {
                     fontFaces = new[]
                     {
-                        new { fontFamily = "Minecrafter-1721628880", fontType = "displayFont", path = "Minecrafter.Reg.ttf" },
-                        new { fontFamily = "Aaargh-1228051400", fontType = "contentFont", path = "Aaargh.ttf" }
+                        new { fontFamily = displayFontUniqueId, fontType = "displayFont", path = displayFont.FileName },
+                        new { fontFamily = contentFontUniqueId, fontType = "contentFont", path = contentFont.FileName }
                     },
                     fontSlots = new
                     {
-                        body = new { fontFace = "Regular", fontFamily = "Minecrafter-1721628880" },
-                        heading = new { fontFace = "Normal", fontFamily = "Aaargh-1228051400" },
-                        label = new { fontFace = "Normal", fontFamily = "Aaargh-1228051400" },
-                        title = new { fontFace = "Regular", fontFamily = "Minecrafter-1721628880" }
+                        body = new { fontFace = bodyFontStyle, fontFamily = bodyFont.Id == displayFont.Id ? displayFontUniqueId : contentFontUniqueId },
+                        heading = new { fontFace = headlineFontStyle, fontFamily = headlineFont.Id == displayFont.Id ? displayFontUniqueId : contentFontUniqueId },
+                        label = new { fontFace = interactiveFontStyle, fontFamily = interactiveFont.Id == displayFont.Id ? displayFontUniqueId : contentFontUniqueId },
+                        title = new { fontFace = titleFontStyle, fontFamily = titleFont.Id == displayFont.Id ? displayFontUniqueId : contentFontUniqueId }
                     }
                 }),
                 IsValid = true,
                 IsHidden = !visble
             });
+
+            return fontpackage;
         }
 
         /// <summary>
@@ -175,6 +202,35 @@ namespace PnP.PowerShell.Commands.Utilities
                 Store.OutOfBox => $"{webUrl}/_api/outofboxfontpackages",
                 _ => throw new ArgumentOutOfRangeException(nameof(store), store, null)
             };
+        }
+
+        /// <summary>
+        /// Retrieves all fonts from the Brand Center fonts library
+        /// </summary>
+        /// <param name="cmdlet">Cmdlet for which this runs, used for logging</param>
+        /// <param name="clientContext">ClientContext to use to communicate with SharePoint Online</param>
+        /// <returns>List with Font instances</returns>
+        public static List<Font> GetFonts(BasePSCmdlet cmdlet, ClientContext clientContext)
+        {
+            var brandCenterConfig = GetBrandCenterConfiguration(cmdlet, clientContext);
+
+            if (brandCenterConfig == null || string.IsNullOrEmpty(brandCenterConfig.SiteUrl) || string.IsNullOrEmpty(brandCenterConfig.BrandFontLibraryUrl.DecodedUrl))
+            {
+                cmdlet?.LogError("Brand Center configuration is not available or incomplete.");
+                return null;
+            }
+
+            cmdlet?.LogDebug("Retrieving all fonts from the Brand Center fonts library at {brandCenterConfig.BrandFontLibraryUrl.DecodedUrl}");
+            var url = $"{brandCenterConfig.SiteUrl}/_api/SP.List.GetListDataAsStream?listFullUrl='{brandCenterConfig.BrandFontLibraryUrl.DecodedUrl}'";
+            // Data is in a Row property
+            var fonts = RestHelper.Post<RestRowCollection<Font>>(Framework.Http.PnPHttpClient.Instance.GetHttpClient(), url, clientContext, new
+            {
+                parameters = new
+                {
+                    ViewXml = "<View><ViewFields><FieldRef Name=\"FileLeafRef\"/><FieldRef Name=\"_SPFontVisible\"/><FieldRef Name=\"_SPFontFamilyName\"/><FieldRef Name=\"_SPFontFaces\"/></ViewFields><Query><Where><And><IsNotNull><FieldRef Name=\"_SPFontFamilyName\"/></IsNotNull><IsNotNull><FieldRef Name=\"_SPFontFaces\"/></IsNotNull></And></Where></Query><RowLimit Paged=\"TRUE\">5000</RowLimit></View>"
+                }
+            });
+            return fonts.Items.ToList();
         }
 
         /// <summary>